+++ springwolf-add-ons/springwolf-generic-binding/src/main/java/io/github/stavshamir/springwolf/asyncapi/addon/genericbinding/annotation/processor/PropertiesUtil.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.asyncapi.scanners.bindings.annotation.processor;
+package io.github.stavshamir.springwolf.asyncapi.addon.genericbinding.annotation.processor;
 
 import lombok.extern.slf4j.Slf4j;
